Standalone Lite: Cluster Environments (#14509)
