+++ src/lightning_lite/plugins/__init__.py
@@ -1,0 +1,1 @@
+from lightning_lite.plugins.environments import ClusterEnvironment  # noqa: F401
