+++ src/lightning_lite/plugins/environments/kubeflow_environment.py
@@ -15,7 +15,7 @@
 import logging
 import os
 
-from pytorch_lightning.plugins.environments.cluster_environment import ClusterEnvironment
+from lightning_lite.plugins.environments.cluster_environment import ClusterEnvironment
 
 log = logging.getLogger(__name__)
