+++ src/lightning_lite/plugins/environments/xla_environment.py
@@ -14,8 +14,8 @@
 import logging
 import os
 
-from pytorch_lightning.plugins.environments.cluster_environment import ClusterEnvironment
-from pytorch_lightning.utilities.imports import _TPU_AVAILABLE
+from lightning_lite.plugins.environments.cluster_environment import ClusterEnvironment
+from lightning_lite.utilities.imports import _TPU_AVAILABLE
 
 if _TPU_AVAILABLE:
     import torch_xla.core.xla_env_vars as xenv
