+++ src/pytorch_lightning/strategies/launchers/subprocess_script.py
@@ -22,7 +22,7 @@ import numpy as np
 from lightning_utilities.core.imports import RequirementCache
 
 import pytorch_lightning as pl
-from pytorch_lightning.plugins.environments.cluster_environment import ClusterEnvironment
+from lightning_lite.plugins.environments.cluster_environment import ClusterEnvironment
 from pytorch_lightning.strategies.launchers.base import _Launcher
 
 _HYDRA_AVAILABLE = RequirementCache("hydra")
