+++ tests/tests_pytorch/lite/test_parity.py
@@ -28,11 +28,11 @@ from torch.nn.parallel.distributed import DistributedDataParallel
 from torch.utils.data import DataLoader
 from torch.utils.data.distributed import DistributedSampler
 
+from lightning_lite.plugins.environments.lightning_environment import find_free_network_port
 from lightning_lite.utilities.apply_func import move_data_to_device
 from lightning_lite.utilities.cloud_io import atomic_save
 from pytorch_lightning.demos.boring_classes import RandomDataset
 from pytorch_lightning.lite import LightningLite
-from pytorch_lightning.plugins.environments.lightning_environment import find_free_network_port
 from pytorch_lightning.strategies.ddp_spawn import DDPSpawnStrategy
 from tests_pytorch.helpers.runif import RunIf
