+++ tests/tests_pytorch/models/test_amp.py
@@ -20,9 +20,9 @@ from torch import optim
 from torch.utils.data import DataLoader
 
 import tests_pytorch.helpers.utils as tutils
+from lightning_lite.plugins.environments import SLURMEnvironment
 from pytorch_lightning import Trainer
 from pytorch_lightning.demos.boring_classes import BoringModel, RandomDataset
-from pytorch_lightning.plugins.environments import SLURMEnvironment
 from tests_pytorch.helpers.runif import RunIf
