+++ tests/tests_pytorch/trainer/connectors/test_signal_connector.py
@@ -19,9 +19,9 @@ from unittest import mock
 
 import pytest
 
+from lightning_lite.plugins.environments import SLURMEnvironment
 from pytorch_lightning import Trainer
 from pytorch_lightning.demos.boring_classes import BoringModel
-from pytorch_lightning.plugins.environments import SLURMEnvironment
 from pytorch_lightning.trainer.connectors.signal_connector import SignalConnector
 from pytorch_lightning.utilities.exceptions import ExitGracefullyException
 from tests_pytorch.helpers.runif import RunIf
